refactor(recipes): split JsonLdRecipeParser into 3 partial files
JsonLdRecipeParser was 272 lines of static helpers covering three
distinct concerns: the entry point + Recipe shape discovery, numeric
field extractors (servings yields, ISO-8601 durations), and content
extractors (ingredients, instruction trees, image URLs).

Split across three partials of the existing partial class:

- JsonLdRecipeParser.cs (139 lines) - TryParse, blob extraction,
  FindRecipe walk, type detection, MapRecipe + the JsonLdScript regex
- JsonLdRecipeParser.Numerics.cs (59 lines) - ReadServings,
  ParseLeadingInt, ReadDurationMinutes + the IsoDuration regex
- JsonLdRecipeParser.Content.cs (92 lines) - ReadIngredients,
  ReadSteps + FlattenInstructions recursion, ReadImageUrl

No public API change; all 181 parser tests pass.
